+++ tests/Support/config/di-web-view-response-with-error.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-use HttpSoft\Message\Response;
-use HttpSoft\Message\StreamFactory;
+use HttpSoft\Message\ResponseFactory;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Message\StreamInterface;
 use Psr\Http\Server\MiddlewareInterface;
 use Psr\Http\Server\RequestHandlerInterface;
-use Yiisoft\Http\Status;
+use Yiisoft\Yii\Runner\Http\Tests\Support\ViewResponse;
 
 return [
     'applicationMiddleware' => new class implements MiddlewareInterface {
@@ -16,9 +16,12 @@ return [
             ServerRequestInterface $request,
             RequestHandlerInterface $handler,
         ): ResponseInterface {
-            return (new Response())
-                ->withBody((new StreamFactory())->createStream('OK'))
-                ->withStatus(Status::CONTINUE);
+            $response = (new ResponseFactory())->createResponse();
+
+            return new ViewResponse(
+                $response,
+                fn(): StreamInterface => throw new Exception('Failure while creating response stream'),
+            );
         }
     },
 ];